requirements.txt
# DataMatrix detection dependencies
pylibdmtx==0.1.10
pillow==11.3.0
pandas==2.3.3
numpy==2.0.2
opencv-python-headless==4.12.0.88